feat: add product metadata to admin billing
